♻️ Extracted DOM utilities from App.js to dom-utils.js
Extracted DOM manipulation and event handler functions:
- getScrollbarWidth() - Calculate scrollbar width for layout
- sendPortalReadyEvent() - Send ready event to parent window
- setupRecommendationButtons() - Setup recommendation click tracking
- showLexicalSignupForms() - Show signup forms when not logged in

Reduced App.js by 68 lines (728 → 660 lines)
All tests passing (256/257)
